services/requirements.txt
# requirements.txt - zhiqu-classroom MVP

# === Web Framework ===
fastapi==0.115.*
uvicorn[standard]==0.34.*
python-multipart==0.0.*

# === Database ===
sqlalchemy[asyncio]==2.0.*
asyncpg==0.30.*
alembic==1.14.*
pgvector==0.3.*

# === Cache & Queue ===
redis[hiredis]==5.2.*

# === Auth ===
pyjwt[crypto]==2.10.*
passlib[bcrypt]==1.7.*

# === Object Storage ===
minio==7.2.*

# === LLM ===
openai==1.61.*
httpx==0.28.*

# === Document Parsing ===
PyMuPDF==1.25.*
pdfplumber==0.11.*
python-docx==1.1.*

# === Utilities ===
pydantic==2.10.*
pydantic-settings==2.7.*
structlog==24.4.*
uuid6==2024.7.*

# === Dev ===
pytest==8.3.*
pytest-asyncio==0.25.*
ruff==0.9.*
